feat: 관리자 재난 제보 그룹 상세 조회 api 추가
- 관리자 재난 제보 글부 상세 조회 추가
- 재난 제보 그룹 등록/수정 관리자 통계 및 아카이브 밑으로 이동
